refactor(components): 重构组件目录结构，将导航和UI组件分组
将导航相关组件(CategoryNav, SearchBar, NavigationButtons)移动到navigation目录
将UI组件(LoadingSpinner, ErrorDisplay)移动到ui目录
更新所有相关导入路径以匹配新的目录结构
