+++ src/components/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as LoadingSpinner } from './LoadingSpinner';
+export { default as ErrorDisplay } from './ErrorDisplay';
